Fix default value defs.
It SEEMS like there's a difference between providing a default value
on /init/ vs on the var def. I don't know why?

I'd assumed they were the same, and for swift Data they DO appear to
be the same, but I'm moving forward under the assumption they make a
difference for cloudkit.