Refactor to use fewer stupid indentions.
For some reason I had a ton of Form/Seciton/Hstack shit, that was
utterly unnecessary.

The UI now looks different on macos and iOS. on MacOS it looks
slightly less like I intended, but I still have plenty of refactoring
to do.

On iOS it now actually works! so, worth it.